+++ crates/client/rpc/Cargo.toml
@@ -17,19 +17,9 @@ version.workspace = true
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-# Deoxys utils
-mp-genesis-config = { workspace = true }
-
-# Deoxys runtime
-pallet-starknet-runtime-api = { workspace = true, default-features = true }
-
 # Deoxys client
 mc-db = { workspace = true }
 mc-sync = { workspace = true }
-
-# Substrate client
-sc-client-api = { workspace = true, default-features = true }
-sc-network-sync = { workspace = true }
 
 # Starknet
 blockifier = { workspace = true, default-features = true }
@@ -50,7 +40,6 @@ jsonrpsee = { workspace = true, default-features = true, features = [
 log = { workspace = true, default-features = true }
 mp-block = { workspace = true, default-features = true }
 mp-convert = { workspace = true, default-features = true }
-mp-digest-log = { workspace = true }
 mp-felt = { workspace = true, default-features = true }
 mp-hashers = { workspace = true, default-features = true }
 mp-simulations = { workspace = true }
