+++ crates/client/rpc/src/methods/read/estimate_message_fee.rs
@@ -3,8 +3,8 @@ use std::sync::Arc;
 use blockifier::transaction::transactions::L1HandlerTransaction;
 use jsonrpsee::core::RpcResult;
 use mp_felt::{Felt252Wrapper, FeltWrapper};
+use mp_hashers::pedersen::PedersenHasher;
 use mp_transactions::compute_hash::ComputeTransactionHash;
-use mp_types::block::DHasherT;
 use starknet_api::core::Nonce;
 use starknet_api::hash::StarkFelt;
 use starknet_api::transaction::{Calldata, Fee, TransactionVersion};
@@ -63,7 +63,7 @@ pub fn convert_message_into_tx(
         calldata: Calldata(Arc::new(calldata)),
     };
     // TODO(merge): recheck if this is correct
-    let tx_hash = tx.compute_hash::<DHasherT>(chain_id, true, block_number);
+    let tx_hash = tx.compute_hash::<PedersenHasher>(chain_id, true, block_number);
 
     L1HandlerTransaction { tx, tx_hash, paid_fee_on_l1: Fee(10) } //TODO: fix with real fee
 }
