+++ crates/node/Cargo.toml
@@ -4,8 +4,7 @@ authors = [
   "KSS <https://github.com/keep-starknet-strange>",
   "Substrate DevHub <https://github.com/substrate-developer-hub>",
 ]
-build = "build.rs"
-description = "Deoxys node."
+description = "Deoxys node"
 edition.workspace = true
 homepage = "https://github.com/kasarlabs/deoxys"
 license = "MIT"
@@ -28,60 +27,61 @@ log = { workspace = true }
 serde = { workspace = true }
 tokio = { workspace = true }
 
-frame-system = { workspace = true }
-sc-basic-authorship = { workspace = true }
-sc-cli = { workspace = true }
-sc-client-api = { workspace = true }
-sc-consensus = { workspace = true }
-sc-consensus-manual-seal = { workspace = true }
-sc-executor = { workspace = true }
-sc-network = { workspace = true }
-sc-network-sync = { workspace = true }
-sc-service = { workspace = true }
-sc-telemetry = { workspace = true }
-sc-transaction-pool = { workspace = true }
-sc-transaction-pool-api = { workspace = true }
-sp-core = { workspace = true }
-sp-inherents = { workspace = true }
-sp-keyring = { workspace = true }
-sp-runtime = { workspace = true }
-sp-state-machine = { workspace = true }
-sp-statement-store = { workspace = true }
-sp-timestamp = { workspace = true }
+# frame-system = { workspace = true }
+# sc-basic-authorship = { workspace = true }
+# sc-cli = { workspace = true }
+# sc-client-api = { workspace = true }
+# sc-consensus = { workspace = true }
+# sc-consensus-manual-seal = { workspace = true }
+# sc-executor = { workspace = true }
+# sc-network = { workspace = true }
+# sc-network-sync = { workspace = true }
+# sc-service = { workspace = true }
+# sc-telemetry = { workspace = true }
+# sc-transaction-pool = { workspace = true }
+# sc-transaction-pool-api = { workspace = true }
+# sp-core = { workspace = true }
+# sp-inherents = { workspace = true }
+# sp-keyring = { workspace = true }
+# sp-runtime = { workspace = true }
+# sp-state-machine = { workspace = true }
+# sp-statement-store = { workspace = true }
+# sp-timestamp = { workspace = true }
 
 # Substrate primitives dependencies
-sp-api = { workspace = true }
-sp-block-builder = { workspace = true }
-sp-blockchain = { workspace = true }
+# sp-api = { workspace = true }
+# sp-block-builder = { workspace = true }
+# sp-blockchain = { workspace = true }
 # Substrate client dependencies
 prometheus-endpoint = { workspace = true }
-sc-rpc-api = { workspace = true }
+# sc-rpc-api = { workspace = true }
 # Substrate frame dependencies
 # no substrate frame pallet dependencies for now
 
 # Substrate tools dependencies
-substrate-frame-rpc-system = { workspace = true }
+# substrate-frame-rpc-system = { workspace = true }
 # These dependencies are used for runtime benchmarking
-frame-benchmarking = { workspace = true }
-frame-benchmarking-cli = { workspace = true }
+# frame-benchmarking = { workspace = true }
+# frame-benchmarking-cli = { workspace = true }
 
 # Starknet
-deoxys-runtime = { workspace = true }
+# deoxys-runtime = { workspace = true }
 mc-db = { workspace = true }
 mc-rpc = { workspace = true }
-pallet-starknet = { workspace = true }
-pallet-starknet-runtime-api = { workspace = true }
+# pallet-starknet = { workspace = true }
+# pallet-starknet-runtime-api = { workspace = true }
 starknet-core = { workspace = true }
 starknet-providers = { workspace = true }
 
 # Deoxys utils
-mc-genesis-data-provider = { workspace = true }
+# mc-genesis-data-provider = { workspace = true }
 
 # Primitives
 mp-block = { workspace = true }
 mp-felt = { workspace = true }
-mp-digest-log = { workspace = true }
+# mp-digest-log = { workspace = true }
 mp-types = { workspace = true }
+primitive-types = { workspace = true }
 
 # CLI-specific dependencies
 try-runtime-cli = { optional = true, git = "https://github.com/kasarlabs/polkadot-sdk", branch = "release-polkadot-v1.3.0-std" }
@@ -105,20 +105,17 @@ tower-http.workspace = true
 serde_json.workspace = true
 env_logger = "0.11.3"
 
-[build-dependencies]
-substrate-build-script-utils = { workspace = true }
-
 [features]
 default = []
 sound = ["mc-sync/m"]
 # Dependencies that are only required if runtime benchmarking should be build.
-runtime-benchmarks = [
-  "frame-benchmarking-cli/runtime-benchmarks",
-  "frame-benchmarking/runtime-benchmarks",
-  "deoxys-runtime/runtime-benchmarks",
-]
+# runtime-benchmarks = [
+#   "frame-benchmarking-cli/runtime-benchmarks",
+#   "frame-benchmarking/runtime-benchmarks",
+#   # "deoxys-runtime/runtime-benchmarks",
+# ]
 # Enable features that allow the runtime to be tried and debugged. Name might be subject to change
 # in the near future.
-disable-transaction-fee = ["deoxys-runtime/disable-transaction-fee"]
-try-runtime = ["deoxys-runtime/try-runtime", "try-runtime-cli/try-runtime"]
+# disable-transaction-fee = ["deoxys-runtime/disable-transaction-fee"]
+# try-runtime = ["deoxys-runtime/try-runtime", "try-runtime-cli/try-runtime"]
 tui = ["deoxys-tui"]
