+++ crates/primitives/block/Cargo.toml
@@ -18,7 +18,6 @@ mp-hashers = { workspace = true }
 mp-transactions = { workspace = true }
 serde = { workspace = true, features = ["derive"], optional = true }
 serde_json = { workspace = true }
-sp-core = { workspace = true }
 starknet-core = { workspace = true }
 starknet_api = { workspace = true }
 
@@ -28,6 +27,7 @@ parity-scale-codec = { workspace = true, features = [
   "derive",
 ], optional = true }
 scale-info = { workspace = true, features = ["derive"], optional = true }
+primitive-types.workspace = true
 
 [dev-dependencies]
 mp-felt = { workspace = true }
@@ -46,7 +46,6 @@ std = [
   "mp-hashers/std",
   "mp-transactions/std",
   "serde/std",
-  "sp-core/std",
   # Optionals
   "anyhow",
   "parity-scale-codec?/std",
