+++ crates/primitives/felt/Cargo.toml
@@ -12,7 +12,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 cairo-vm = { workspace = true }
-sp-core = { workspace = true }
+primitive-types.workspace = true
 starknet-core = { workspace = true }
 starknet-ff = { workspace = true }
 starknet-types-core = { workspace = true }
@@ -36,7 +36,6 @@ serde = ["dep:serde", "starknet-ff/serde", "dep:serde_with"]
 std = [
   "cairo-vm/std",
   "thiserror-no-std/std",
-  "sp-core/std",
   "starknet-ff/std",
   "starknet-core/std",
   # Optional
