+++ crates/primitives/felt/src/lib.rs
@@ -26,7 +26,7 @@ pub use convert::*;
 use parity_scale_codec::{Decode, Encode, EncodeLike, Error, Input, MaxEncodedLen, Output};
 #[cfg(feature = "scale-info")]
 use scale_info::{build::Fields, Path, Type, TypeInfo};
-use sp_core::{H256, U256};
+use primitive_types::{H256, U256};
 use starknet_api::hash::StarkFelt;
 use starknet_ff::{FieldElement, FromByteSliceError, FromStrError};
 use starknet_types_core::felt::Felt;
