+++ crates/primitives/genesis-config/src/lib.rs
@@ -12,22 +12,6 @@ use serde_with::serde_as;
 use starknet_core::serde::unsigned_field_element::UfeHex;
 use starknet_crypto::FieldElement;
 
-// 0x49d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7
-pub const ETH_TOKEN_ADDR: HexFelt = HexFelt(FieldElement::from_mont([
-    4380532846569209554,
-    17839402928228694863,
-    17240401758547432026,
-    418961398025637529,
-]));
-
-// 0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d
-pub const STRK_TOKEN_ADDR: HexFelt = HexFelt(FieldElement::from_mont([
-    16432072983745651214,
-    1325769094487018516,
-    5134018303144032807,
-    468300854463065062,
-]));
-
 /// A wrapper for FieldElement that implements serde's Serialize and Deserialize for hex strings.
 #[serde_as]
 #[derive(Serialize, Deserialize, Copy, Clone)]
