+++ crates/primitives/types/src/block.rs
@@ -1,23 +1,14 @@
-use mp_hashers::pedersen::PedersenHasher;
-use sp_core::H256;
-use sp_runtime::generic::{Block, BlockId, Header};
-use sp_runtime::traits::BlakeTwo256;
-use sp_runtime::OpaqueExtrinsic;
+// use mp_hashers::pedersen::PedersenHasher;
+use primitive_types::H256;
 
 // Deoxys block number type
-pub type DBlockNumber = u32;
+// pub type DBlockNumber = u32;
 
-// Deoxys block header type
-pub type DHeaderT = Header<DBlockNumber, BlakeTwo256>;
-
-// Deoxys block type
-pub type DBlockT = Block<DHeaderT, OpaqueExtrinsic>;
-
-// Deoxys hash type
+// // Deoxys hash type
 pub type DHashT = H256;
 
-// Deoxys default hasher type
-pub type DHasherT = PedersenHasher;
+// // Deoxys default hasher type
+// pub type DHasherT = PedersenHasher;
 
-// Deoxys block identifier type
-pub type DBlockID = BlockId<DBlockT>;
+// // Deoxys block identifier type
+// pub type DBlockID = BlockId<DBlockT>;
